+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 Unreleased
 ------------------------
 - Enh #69: Add `requirements.php`
+- Fix #71: After module disabling, Content for ExternalCalendarEntry are not deleted
 
 1.4.2 (January 19, 2023)
 -----------------------