+++ docs/CHANGELOG.md
@@ -1,8 +1,11 @@
 Changelog
 =========
 
-TBA
+Unreleased
+------------------------
 - Enh #68: Make URL locations clickable
+- Enh #69: Add `requirements.php`
+- Fix #71: After module disabling, Content for ExternalCalendarEntry are not deleted
 
 1.4.2 (January 19, 2023)
 -----------------------